backend/requirements.txt
fastapi
uvicorn
google-cloud-vision
gtts
numpy
opencv-python
pillow